Better cursor handling
This commit completely changes the way cursor is used in the ``Database``
class. In the previous versions of squery-lite, the cursor was always resused
except in transactions. This did not allow us to use the iterator feature of
the cursor object which could become a problem when reduced memory footprint is
desired.

In this patch, the cursor is now *never* reused. For each of the
query-executing methods, such as query, execute, executemany, etc, a new cursor
is created and returned.

A new wrapper class, ``Cursor`` is introduced to keep the end user code change
to minimum, but significant refactoring is still needed. One example of the way
user code will change is this::

    db = squery.Database(connection)
    db.query(query, arg)
    results = db.results

will change to::

    db = squery.Database(connection)
    cursor = db.query(query, arg)
    results = cursor.results

The cursor object returned by query-executing methods on the Database objects
are iterable. This allows writing efficient generator functions and methods::

    db = squery.Database(connection)
    cursor = db.query(query, arg)
    for result in cursor:
        yield MyClass(result)

Signed-off-by: Branko Vukelic <branko@outernet.is>