Give transaction ability to start a new connection
This commit modifies the transaction method so that it can initiate a new
database connection. This feature is added to enable two possible use cases:

- better isolation btween transactions
- ability to register custom aggregates and functions *per transaction*

In order to implement this feature, the Connection class has been modified with
a new() method which returns a new instance of the connection object to the
same database as the original object.

Additionally, the Cursor class has been modified to accept the connection
object instead of a sqlite3 cursor instance. This allows the user to gain
access to whatever connection is being used by the cursor (and, by extension,
by the transaction).

The transaction() method now has a new_connection argument, which is False by
default, and starts a new connection when it is True. For example::

    with db.transaction(new_connection=True) as cur:
        # Not using the same connection
        assert cur.conn is not db.conn

Signed-off-by: Branko Vukelic <branko@outernet.is>